desktop: added wallpaperpacks support.
A wallpaperpack is a simple, ConfigParser-style, configuration file that
specifies some informations on the wallpapers in a directory ("pack").
If the selected wallpaper has these extra informations, a button in the middle
of the toolbar will appear, and by clicking on it the user will open a new dialog
with the requested informations.