ApplicationSelectionDialog: do not destroy at delete-event.
Re-generating the dialog everytime the user needs it is not great because iterating
on GMenu is not fast, even if the items are already in cache.
When delete-event is fired, the dialog is just hidden.

It's up to modules to destroy the dialog when they do not need it anymore, for example
when going back to the main scene.